Fix for timer update interrupt not activating ADC
When creating a Timer wrapper object, the registers associated with the
given TIM# object are reset to defaults. In order to ensure the timer
update interrupt it set an unsafe reference to the TIM# must be created
in order to set the MMS register.

Using this information the timer-interrupt-adc-hal.rs program can
configure a PWM on TIM2 and a ADC triggered by updated on TIM2, while
still using the stm32f4xx-hal to configure the PWM and ADC.